+++ code/EleicoesTransparentesPersistance/src/br/ufba/mata62/eleicoestransparentes/model/logicbeans/PontosGrafico.java
@@ -1,4 +1,4 @@
-package br.ufba.mata62.eleicoestransparentes.business;
+package br.ufba.mata62.eleicoestransparentes.model.logicbeans;
 
 import java.util.ArrayList;
