+++ code/EleicoesTransparentesPersistance/src/ormliteLocalLog.properties
@@ -1,0 +1,4 @@
+com\.j256\.ormlite.* = INFO 
+com\.j256\.ormlite\.stmt\.mapped\.BaseMappedStatement = TRACE
+com\.j256\.ormlite\.stmt\.mapped\.MappedCreate = TRACE
+com\.j256\.ormlite\.stmt\.StatementExecutor = TRACE
