+++ code/ParserTSE/src/br/ufba/mata62/eleicoestransparentes/business/perfil/PerfilPartido.java
@@ -1,4 +1,4 @@
-package br.ufba.mata62.eleicoestransparentes.business;
+package br.ufba.mata62.eleicoestransparentes.business.perfil;
 
 import java.text.Normalizer;
 import java.util.regex.Matcher;
