+++ code/RadarEleitoral/src/br/ufba/mata62/eleicoestransparentes/view/chart/GraficoBarraDoadores.java
@@ -42,11 +42,10 @@
  *
  */
 
-package br.ufba.mata62.eleicoestransparentes.view;
+package br.ufba.mata62.eleicoestransparentes.view.chart;
 
 import java.awt.Color;
 import java.awt.Dimension;
-import java.awt.Font;
 import java.awt.GradientPaint;
 import java.util.ArrayList;
 
@@ -57,102 +56,60 @@ import org.jfree.chart.axis.CategoryAxis;
 import org.jfree.chart.axis.CategoryLabelPositions;
 import org.jfree.chart.axis.NumberAxis;
 import org.jfree.chart.plot.CategoryPlot;
-import org.jfree.chart.plot.PiePlot;
 import org.jfree.chart.plot.PlotOrientation;
 import org.jfree.chart.renderer.category.BarRenderer;
 import org.jfree.data.category.CategoryDataset;
 import org.jfree.data.category.DefaultCategoryDataset;
-import org.jfree.data.general.DefaultPieDataset;
-import org.jfree.data.general.PieDataset;
 import org.jfree.ui.ApplicationFrame;
 import org.jfree.ui.RefineryUtilities;
 
-
-import javax.swing.JPanel;
-
-import org.jfree.chart.ChartFactory;
-import org.jfree.chart.ChartPanel;
-import org.jfree.chart.JFreeChart;
-import org.jfree.chart.plot.PiePlot;
-import org.jfree.data.general.DefaultPieDataset;
-import org.jfree.data.general.PieDataset;
-import org.jfree.ui.ApplicationFrame;
-import org.jfree.ui.RefineryUtilities;
-
-import br.ufba.mata62.eleicoestransparentes.business.GraficoFinal;
-import br.ufba.mata62.eleicoestransparentes.business.PontosGrafico;
+import br.ufba.mata62.eleicoestransparentes.business.grafico.GraficoFinal;
+import br.ufba.mata62.eleicoestransparentes.model.logicbeans.PontosGrafico;
 
 /**
  * A simple demonstration application showing how to create a bar chart.
  *
  */
-public class GraficoPizzaCandidatos extends ApplicationFrame {
+public class GraficoBarraDoadores extends ApplicationFrame {
 
     /**
 	 * 
 	 */
-	
-    public GraficoPizzaCandidatos(GraficoFinal graficoFinal) {
+	private static final long serialVersionUID = 1L;
+
+	public GraficoBarraDoadores(GraficoFinal graficoFinal) {
+
         super(graficoFinal.getNome());
-        setContentPane(createDemoPanel(graficoFinal));
-    }
 
-    /**
-     * Creates a sample dataset.
-     * 
-     * @return A sample dataset.
-     */
-    private static PieDataset createDataset(ArrayList<PontosGrafico> pontos) {
-        DefaultPieDataset dataset = new DefaultPieDataset();
-       
-        for(PontosGrafico ponto: pontos)
-        	dataset.setValue(ponto.rotulo, new Double(ponto.valor));
-        
-        return dataset;        
-    }
-    
-    /**
-     * Creates a chart.
-     * 
-     * @param dataset  the dataset.
-     * 
-     * @return A chart.
-     */
-    private static JFreeChart createChart(PieDataset dataset, String titulo) {
-        
-        JFreeChart chart = ChartFactory.createPieChart(
-            titulo,  // chart title
-            dataset,             // data
-            true,               // include legend
-            true,
-            false
-        );
+        final CategoryDataset dataset = createDataset(graficoFinal.getPontosGrafico());
+        final JFreeChart chart = createChart(dataset, graficoFinal.getNome());
+        final ChartPanel chartPanel = new ChartPanel(chart);
+        chartPanel.setPreferredSize(new Dimension(500, 270));
+        setContentPane(chartPanel);
 
-        PiePlot plot = (PiePlot) chart.getPlot();
-        plot.setLabelFont(new Font("SansSerif", Font.PLAIN, 12));
-        plot.setNoDataMessage("Nenhum dado disponibilizado");
-        plot.setCircular(false);
-        plot.setLabelGap(0.02);
-        return chart;
-        
     }
-    
-    /**
-     * Creates a panel for the demo (used by SuperDemo.java).
-     * 
-     * @return A panel.
-     */
-    public static JPanel createDemoPanel(GraficoFinal graficoFinal) {
-        JFreeChart chart = createChart(createDataset(graficoFinal.getPontosGrafico()), graficoFinal.getNome());
-        return new ChartPanel(chart);
-    }
-	
 
     /**
      * Returns a sample dataset.
      * 
      * @return The dataset.
      */
+    private CategoryDataset createDataset(ArrayList<PontosGrafico> pontos) {
+        
+        // row keys...
+        final String series1 = "";
+
+        // column keys...
+
+        // create the dataset...
+        final DefaultCategoryDataset dataset = new DefaultCategoryDataset();
+        
+        for(PontosGrafico ponto: pontos)
+        dataset.addValue(ponto.valor, series1, ponto.rotulo);
+
+        return dataset;
+        
+    }
     
     /**
      * Creates a sample chart.
@@ -161,7 +118,55 @@ public class GraficoPizzaCandidatos extends ApplicationFrame {
      * 
      * @return The chart.
      */
+    private JFreeChart createChart(final CategoryDataset dataset, String titulo) {
+        
+        // create the chart...
+        final JFreeChart chart = ChartFactory.createBarChart(
+            titulo,                 // chart title
+            "Doador",               // domain axis label
+            "Valor",                  // range axis label
+            dataset,                  // data
+            PlotOrientation.VERTICAL, // orientation
+            true,                     // include legend
+            true,                     // tooltips?
+            false                     // URLs?
+        );
 
+        // NOW DO SOME OPTIONAL CUSTOMISATION OF THE CHART...
+
+        // set the background color for the chart...
+        chart.setBackgroundPaint(Color.white);
+
+        // get a reference to the plot for further customisation...
+        final CategoryPlot plot = chart.getCategoryPlot();
+        plot.setBackgroundPaint(Color.lightGray);
+        plot.setDomainGridlinePaint(Color.white);
+        plot.setRangeGridlinePaint(Color.white);
+
+        // set the range axis to display integers only...
+        final NumberAxis rangeAxis = (NumberAxis) plot.getRangeAxis();
+        rangeAxis.setStandardTickUnits(NumberAxis.createIntegerTickUnits());
+
+        // disable bar outlines...
+        final BarRenderer renderer = (BarRenderer) plot.getRenderer();
+        renderer.setDrawBarOutline(false);
+        
+        // set up gradient paints for series...
+        final GradientPaint gp0 = new GradientPaint(
+            0.0f, 0.0f, Color.blue, 
+            0.0f, 0.0f, Color.lightGray
+        );
+        renderer.setSeriesPaint(0, gp0);
+
+        final CategoryAxis domainAxis = plot.getDomainAxis();
+        domainAxis.setCategoryLabelPositions(
+            CategoryLabelPositions.createUpRotationLabelPositions(Math.PI / 6.0)
+        );
+        // OPTIONAL CUSTOMISATION COMPLETED.
+        
+        return chart;
+        
+    }
     
     // ****************************************************************************
     // * JFREECHART DEVELOPER GUIDE                                               *
